Custom fonts, and some minor fixes. However, seg fault in unit test now..

+++ software/firmware_workspace/tm4c123_freertos_comscicalc/fonts/font.h
@@ -80,7 +80,7 @@ typedef struct font {
     /**
      * @param fontTableSize Size of the font table in bytes
      */
-    uint16_t fontTableSize;
+    uint32_t fontTableSize;
 }font_t;
